Backend/Backend/app/services/ai_validation_service.py
"""
AI validation service for sunglasses detection
"""
import io
import base64
import logging
from typing import Dict, Any
from PIL import Image
from app.core.config import settings

logger = logging.getLogger(__name__)


class AIValidationService:
    """AI validation service for sunglasses detection"""
    
    def __init__(self):
        self.confidence_threshold = settings.AI_CONFIDENCE_THRESHOLD
    
    async def validate_sunglasses(self, image_content: bytes) -> Dict[str, Any]:
        """Validate if image contains sunglasses"""
        try:
            # For now, use the existing logic from main.py
            # This can be refactored to use the AI model or Vision API
            result = self._analyze_image_with_ai_model(image_content)
            return result
        except Exception as e:
            logger.error(f"AI validation failed: {e}")
            raise
    
    def _analyze_image_with_ai_model(self, image_content: bytes) -> Dict[str, Any]:
        """Analyze image using AI model (placeholder implementation)"""
        # This is a simplified version of the existing logic
        # In a real implementation, you would use the actual AI model
        
        try:
            # Load and analyze the image
            image = Image.open(io.BytesIO(image_content))
            width, height = image.size
            
            # Calculate basic statistics
            gray_image = image.convert('L')
            pixels = list(gray_image.getdata())
            total_pixels = len(pixels)
            mean_brightness = sum(pixels) / total_pixels if total_pixels > 0 else 128
            
            # Simple heuristic for sunglasses detection
            dark_threshold = mean_brightness * 0.7
            dark_pixels = sum(1 for pixel in pixels if pixel < dark_threshold)
            dark_ratio = dark_pixels / total_pixels if total_pixels > 0 else 0
            
            # Calculate confidence
            confidence = min(0.9, dark_ratio * 2) if dark_ratio > 0.1 else 0.2
            
            # Determine if sunglasses are detected
            is_sunglasses = confidence > self.confidence_threshold
            
            # Create response
            status = "accepted" if is_sunglasses else "rejected"
            message = f"Sunglasses detected with {confidence:.1%} confidence" if is_sunglasses else "No sunglasses found in the image"
            
            return {
                "status": status,
                "confidence": confidence,
                "message": message,
                "details": f"Analysis result: {message}",
                "analysis": {
                    "sunglasses_detected": is_sunglasses,
                    "confidence": confidence,
                    "objects": [{
                        "object": "Sunglasses" if is_sunglasses else "No sunglasses",
                        "confidence": confidence,
                        "bounding_box": {"x": 0.2, "y": 0.2, "width": 0.6, "height": 0.3}
                    }] if is_sunglasses else [],
                    "labels": [{
                        "label": "Sunglasses" if is_sunglasses else "No eyewear detected",
                        "confidence": confidence
                    }],
                    "analysis_method": "simplified_ai_model"
                },
                "timestamp": "2024-01-01T00:00:00Z"
            }
            
        except Exception as e:
            logger.error(f"Image analysis failed: {e}")
            # Fallback response
            return {
                "status": "rejected",
                "confidence": 0.0,
                "message": "Image analysis failed",
                "details": f"Error: {str(e)}",
                "analysis": {
                    "sunglasses_detected": False,
                    "confidence": 0.0,
                    "objects": [],
                    "labels": [],
                    "analysis_method": "error_fallback"
                },
                "timestamp": "2024-01-01T00:00:00Z"
            }
