Backend/Backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-strip-extras --output-file=requirements.txt requirements.in
#
absl-py==2.3.1
    # via tink
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   starlette
bazel-runfiles==1.7.0
    # via tink
bcrypt==5.0.0
    # via passlib
cachetools==5.5.2
    # via google-auth
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via uvicorn
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
email-validator==2.3.0
    # via pydantic
fastapi==0.104.1
    # via -r requirements.in
google-api-core==2.28.1
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.123.0
    # via -r requirements.in
google-auth==2.23.4
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
    #   google-genai
google-auth-httplib2==0.1.1
    # via google-api-python-client
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==2.14.0
    # via -r requirements.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-genai==1.2.0
    # via -r requirements.in
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.72.0
    # via google-api-core
greenlet==3.2.4
    # via sqlalchemy
gunicorn==21.2.0
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   requests
packaging==25.0
    # via gunicorn
passlib[bcrypt]==1.7.4
    # via -r requirements.in
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   tink
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic[email]==2.9.0
    # via
    #   -r requirements.in
    #   fastapi
    #   google-genai
pydantic-core==2.23.2
    # via pydantic
pyparsing==3.2.5
    # via httplib2
python-dotenv==1.0.0
    # via -r requirements.in
python-jose==3.3.0
    # via -r requirements.in
requests==2.31.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   google-genai
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.36
    # via -r requirements.in
starlette==0.27.0
    # via fastapi
tink==1.13.0
    # via -r requirements.in
typing-extensions==4.15.0
    # via
    #   fastapi
    #   google-genai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
tzdata==2025.2
    # via pydantic
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn==0.24.0
    # via -r requirements.in
websockets==14.2
    # via google-genai
